Add files via upload
This project aims to predict bankrupt companies using a Random Forest Classifier. The dataset used for this project contains information about various companies, including their financial and non-financial characteristics, as well as whether or not they went bankrupt within a certain period of time. The goal of this project is to build a machine learning model that can accurately predict whether a company will go bankrupt based on its features.